refactor(services): extract web skill catalogue, bindings and overlay
Moves 33 methods out of the composition root into three modules over one
skill-service.ts interface file, whose signatures move out of
AgentService rather than being copied. getSkillOverview declares
`this: AgentService` so its calls to listSkills and detectSkillDrift stay
late-bound through the composed object.

The 14 overlay delegations were blocked in split-web-agent-client because
the runtime's resolver closed over the webSkills binding. It now reads
the catalogue through the state module's accessor, so the runtime
construction moves with the delegations.

web-agent-client.ts 3502 -> 2973, agent-service.ts 496 -> 422.
